+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+[Unreleased]: https://github.com/gg-scm/gg-git/compare/v0.6.0...main
+
+### Added
+
+-  The `LogOptions.NoWalk` field permits reading commits in bulk without having
+   to process their ancestors. ([#2](https://github.com/gg-scm/gg-git/issues/2))
+
 ## [0.6.0][] - 2020-09-27
 
 Version 0.6 introduced an interface for supplying your own Git process creation