+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.7.0][] - 2020-10-02
 
-[Unreleased]: https://github.com/gg-scm/gg-git/compare/v0.6.0...main
+Version 0.7 made improvements to fetching commit information.
+
+[0.7.0]: https://github.com/gg-scm/gg-git/releases/tag/v0.7.0
 
 ### Added